+++ app/src/main/java/com/example/banca/domain/utils/PdfExporter.kt
@@ -11,7 +11,11 @@ object PdfExporter {
     fun exportarLista(
         context: Context,
         titulo: String,
-        elementos: List<String>
+        elementos: List<String>,
+        totalJugado: Double,
+        premios: Double,
+        banco: Double,
+        listero: Double
     ): File {
 
         val document = PdfDocument()
@@ -30,16 +34,40 @@ object PdfExporter {
         paint.textSize = 14f
         paint.isFakeBoldText = false
 
-        var y = 100f
+        var y = 90f
+
+        canvas.drawText("Resultados Lotto", 40f, y, paint)
+        y += 30f
+
+        canvas.drawText("Pick3: 260", 40f, y, paint)
+        y += 25f
+        canvas.drawText("Pick4: 2221", 40f, y, paint)
+        y += 35f
+
+        canvas.drawText("Total jugado: ${"%.0f".format(totalJugado)}", 40f, y, paint)
+        y += 25f
+        canvas.drawText("Premios: ${"%.0f".format(premios)}", 40f, y, paint)
+        y += 25f
+        canvas.drawText("Banco: ${"%.0f".format(banco)}", 40f, y, paint)
+        y += 25f
+        canvas.drawText("Listero: ${"%.0f".format(listero)}", 40f, y, paint)
+        y += 40f
+
+        canvas.drawText("Jugadas:", 40f, y, paint)
+        y += 30f
 
         elementos.forEachIndexed { index, item ->
             canvas.drawText("${index + 1}. $item", 40f, y, paint)
-            y += 30f
+            y += 25f
         }
 
         document.finishPage(page)
 
-        val file = File(context.cacheDir, "$titulo.pdf")
+        val file = File(
+            context.getExternalFilesDir(null),
+            "$titulo.pdf"
+        )
+
         document.writeTo(FileOutputStream(file))
         document.close()